+++ NpcAdventure/Loader/ContentPacks/Provider/LegacyDataProvider.cs
@@ -26,6 +26,7 @@ namespace NpcAdventure.Loader.ContentPacks.Provider
 
             patches.AddRange(this.GetPatchesForAsset(path, "Replace"));
             patches.AddRange(this.GetPatchesForAsset(path, "Patch"));
+            patches.AddRange(this.GetTranslationPatches(path, this.Managed.Pack.Translation.Locale?.ToLower()));
 
             if (patches.Count() < 1)
             {
@@ -34,7 +35,7 @@ namespace NpcAdventure.Loader.ContentPacks.Provider
 
             foreach (var patch in patches)
             {
-                if (patch.Action == "Replace" && this.Managed.Contents.AllowUnsafePatches)
+                if (patch.Action == "Replace")
                 {
                     if (target.Count > 0)
                         this.Monitor.Log(
@@ -45,30 +46,9 @@ namespace NpcAdventure.Loader.ContentPacks.Provider
 
                 var isLocalized = !string.IsNullOrEmpty(patch.Locale);
                 var patchData = this.Managed.Pack.LoadAsset<Dictionary<TKey, TValue>>(patch.FromFile);
-                var possiblyOverrided = AssetPatchHelper.ApplyPatch(target, patchData, patch.CanOverride || isLocalized);
-
-                this.Monitor.Log($"Applied content patch `{patch.LogName}` from content pack `{contentPackName}`");
-
-                if (isLocalized && patchData.Count() > 0)
-                {
-                    var covered = possiblyOverrided.Count();
-                    var extraAdded = patchData.Count() - covered;
-
-                    this.Monitor.Log($"Applied content pack `{contentPackName}` translation patch `{patch.LogName}` to `{patch.Locale}` for `{path}`, covered {covered} keys ({extraAdded} extra added)");
-                }
-                else if (possiblyOverrided.Count() > 0)
-                {
-                    var loglevel = (this.paranoid || !patch.CanOverride) ? LogLevel.Alert : LogLevel.Trace;
-                    this.Monitor.Log($"Found content data key conflicts for `{patch.Target}` by patch `{patch.LogName}` in content pack `{contentPackName}`.", loglevel);
-                    this.Monitor.Log($"   Conflicted keys: {string.Join(", ", possiblyOverrided)}", loglevel);
-                    this.Monitor.Log($"Affected parts of contents {(patch.CanOverride ? "ARE OVERRIDDEN!!!" : "are NOT overridden.")}", loglevel);
-
-                    if (!patch.CanOverride)
-                    {
-                        this.Monitor.Log(
-                            $"If you want to override these keys, you can allow it by set `CanOverride: true` on this patch and allow unsafe patches for this content pack by `AllowUnsafePatches: true`", loglevel);
-                    }
-                }
+                
+                AssetPatchHelper.ApplyPatch(target, patchData);
+                this.Monitor.Log($"Content pack `{contentPackName}` applied{(isLocalized ? $" `{patch.Locale}` translation" : "")} patch `{patch.LogName}` for `{path}`");
             }
 
             return true;
@@ -76,19 +56,18 @@ namespace NpcAdventure.Loader.ContentPacks.Provider
 
         private List<LegacyChanges> GetPatchesForAsset(string path, string action)
         {
-            var locale = this.Managed.Pack.Translation.Locale?.ToLower();
-            var patches = this.Managed.Contents.Changes
+            return this.Managed.Contents.Changes
                 .Where((p) => p.Action.Equals(action) && p.Target.Equals(path) && !p.Disabled)
-                .Where((p) => string.IsNullOrEmpty(p.Locale) || p.Locale.ToLower().Equals(locale))
+                .Where((p) => string.IsNullOrEmpty(p.Locale))
                 .ToList();
+        }
 
-            patches.Sort((a, b) => {
-                if (string.IsNullOrEmpty(a.Locale) && !string.IsNullOrEmpty(b.Locale)) return -1;
-                else if (!string.IsNullOrEmpty(a.Locale)) return 1;
-                return 0;
-            });
-
-            return patches;
+        private List<LegacyChanges> GetTranslationPatches(string path, string locale)
+        {
+            return this.Managed.Contents.Changes
+                .Where((p) => p.Action.Equals("Patch") && p.Target.Equals(path) && !p.Disabled)
+                .Where((p) => !string.IsNullOrEmpty(p.Locale) && p.Locale.ToLower().Equals(locale))
+                .ToList();
         }
     }
 }